% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_multiplegt_dyn.R
\name{did_multiplegt_dyn}
\alias{did_multiplegt_dyn}
\title{Core function for did_multiplegt_dyn}
\usage{
did_multiplegt_dyn(
  df,
  Y,
  G,
  T,
  D,
  effects = 1,
  placebo = 0,
  ci_level = 95,
  switchers = "",
  trends_nonparam = NULL,
  weight = NULL,
  controls = NULL,
  dont_drop_larger_lower = FALSE,
  save_sample = FALSE,
  drop_if_d_miss_before_first_switch = FALSE,
  cluster = NULL,
  same_switchers = FALSE,
  same_switchers_pl = FALSE,
  effects_equal = FALSE,
  save_results = NULL,
  normalized = FALSE,
  design = NULL,
  date_first_switch = NULL,
  normalized_weights = NULL,
  graph_off = FALSE,
  by = NULL,
  predict_het = NULL
)
}
\arguments{
\item{df}{Dataset}

\item{Y}{Outcome}

\item{G}{Group}

\item{T}{Time}

\item{D}{Treatment}

\item{effects}{N. of effects to be computed}

\item{placebo}{N. of placebos}

\item{ci_level}{Confidence level}

\item{switchers}{Switcher option}

\item{trends_nonparam}{Trends_nonparam}

\item{weight}{Weights}

\item{controls}{controls}

\item{dont_drop_larger_lower}{dont_drop_larger_lower}

\item{save_sample}{save_sample}

\item{drop_if_d_miss_before_first_switch}{drop_if_d_miss_before_first_switch}

\item{cluster}{cluster}

\item{same_switchers}{same_switchers}

\item{same_switchers_pl}{same_switchers_pl}

\item{effects_equal}{effects_equal}

\item{save_results}{save_results}

\item{normalized}{normalized}

\item{design}{design}

\item{date_first_switch}{date_first_switch}

\item{normalized_weights}{normalized_weights}

\item{graph_off}{graph_off}

\item{by}{by}

\item{predict_het}{predict_het}
}
\description{
Core function for did_multiplegt_dyn
}
