% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.did_multiplegt_dyn}
\alias{print.did_multiplegt_dyn}
\title{A print method for did_multiplegt_dyn}
\usage{
\method{print}{did_multiplegt_dyn}(x, ...)
}
\arguments{
\item{x}{A did_multiplegt_dyn object}

\item{...}{Undocumented}
}
\value{
No return, custom print method for did_multiplegt_dyn objects. Estimation tables are fetched from the object and displayed in the same style as the Stata did_multiplegt_dyn command.
}
\description{
A customized printed display for did_multiplegt_dyn output
}
