% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnames.R
\name{rnames.did_multiplegt_dyn}
\alias{rnames.did_multiplegt_dyn}
\title{rnames method for did_multiplegt_dyn}
\usage{
\method{rnames}{did_multiplegt_dyn}(obj, ignore = c("plot", "args"), ...)
}
\arguments{
\item{obj}{A did_multiplegt_dyn object}

\item{ignore}{Sublists to be ignored}

\item{...}{Undocumented}
}
\value{
The same output as rnames.
}
\description{
A customized rnames method for did_multiplegt_dyn output
}
